+++ code/core/src/basis_elements/HUDElement.java
@@ -1,25 +1,20 @@
-package interfaces;
+package basis_elements;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import graphic.Painter;
+import graphic.HUDPainter;
 import tools.Point;
 
-public abstract class Entity {
-    private Painter painter;
+public abstract class HUDElement {
     private SpriteBatch batch;
+    private HUDPainter painter;
 
-    /**
-     * Must be implemented for all objects that should be controlled by the <code>EntityController
-     * </code>.
-     */
-    public Entity(Painter painter, SpriteBatch batch) {
-        this.painter = painter;
+    public HUDElement(HUDPainter painter, SpriteBatch batch) {
         this.batch = batch;
+        this.painter = painter;
     }
 
     /** Will be executed every frame. */
     public void update() {}
-    ;
 
     /** @return <code>true</code>, if this instance can be deleted; <code>false</code> otherwise */
     public boolean removable() {
@@ -29,6 +24,7 @@ public abstract class Entity {
     public SpriteBatch getBatch() {
         return batch;
     }
+    ;
 
     /** @return the exact position in the dungeon of this instance */
     public abstract Point getPosition();
@@ -37,12 +33,22 @@ public abstract class Entity {
     public abstract String getTexture();
 
     /** Each drawable should use this <code>Painter</code> to draw itself. */
-    public Painter getPainter() {
+    public HUDPainter getPainter() {
         return painter;
     }
 
     /** Draws this instance on the batch. */
     public void draw() {
         getPainter().draw(getTexture(), getPosition(), getBatch());
+    }
+
+    /**
+     * Draws this instance on the batch with a scaling.
+     *
+     * @param xScaling x-scale
+     * @param yScaling y-scale
+     */
+    public void drawWithScaling(float xScaling, float yScaling) {
+        getPainter().drawWithScaling(xScaling, yScaling, getTexture(), getPosition(), getBatch());
     }
 }
