+++ code/core/src/controller/EntityController.java
@@ -1,6 +1,6 @@
 package controller;
 
-import interfaces.Entity;
+import basis_elements.Entity;
 
 /** Keeps a set of entities and calls their update method every frame. */
 public class EntityController extends AbstractController<Entity> {
