+++ code/core/test/controller/HUDControllerTest.java
@@ -11,7 +11,7 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.freetype.FreeTypeFontGenerator;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
-import interfaces.HUDElement;
+import basis_elements.HUDElement;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
